Extract params and props handling to separate file
