+++ packages/astro/src/core/render/index.ts
@@ -1,11 +1,11 @@
 export { createRenderContext } from './context.js';
 export type { RenderContext } from './context.js';
+export { renderPage } from './core.js';
+export type { Environment } from './environment';
+export { createBasicEnvironment, createEnvironment } from './environment.js';
 export {
 	getParamsAndProps,
 	GetParamsAndPropsError,
 	getParamsAndPropsOrThrow,
-	renderPage,
-} from './core.js';
-export type { Environment } from './environment';
-export { createBasicEnvironment, createEnvironment } from './environment.js';
+} from './params-and-props.js';
 export { loadRenderer, loadRenderers } from './renderer.js';
